Cleaning up new factor!
Also, I think the padding of RHS is a bug---since we only care about
subsequences the same length as @rules[nonterm], we don't want each_cons
to start looking at shorter subsequences padded by spaces.  In
particular, if some suffix of @rules[target] plus the padding happens to
match @rules[nonterm], we'd be in trouble.